requirements.txt
Django
djangorestframework
requests
django-environ
python-decouple